Reorg file structure * Moving all pomodoro code into its own module for separation of   concern
